Created an object file for objects

+++ Sharma_Shubham.assignment-1.08/dims.h
@@ -10,6 +10,6 @@ typedef enum dim {
   num_dims
 } dim_t;
 
-typedef int16_t pair_t[num_dims];
+typedef int8_t pair_t[num_dims];
 
 #endif
